fix: 재고 집계 필드 타입 변경 (int32 -> int64)
통합 집계 시 Integer.MAX_VALUE 오버플로우 위험 방지

+++ services/wms-service/src/main/java/com/kurly/wms/presentation/dto/ProductInventorySummaryResponse.java
@@ -8,13 +8,13 @@ public record ProductInventorySummaryResponse(
         /** 조회 시 warehouseId를 지정하지 않았으면(전국 통합 집계) null이다. */
         Long warehouseId,
 
-        Integer totalQuantity,
-        Integer totalReservedQuantity,
-        Integer totalAvailableQuantity
+        Long totalQuantity,
+        Long totalReservedQuantity,
+        Long totalAvailableQuantity
 ) {
     public static ProductInventorySummaryResponse of(ProductInventorySummaryProjection projection, Long warehouseId) {
-        int totalQuantity = projection.totalQuantity().intValue();
-        int totalReservedQuantity = projection.totalReservedQuantity().intValue();
+        long totalQuantity = projection.totalQuantity();
+        long totalReservedQuantity = projection.totalReservedQuantity();
         return new ProductInventorySummaryResponse(
                 projection.productId(),
                 warehouseId,
